extract gossip functions into own module
